move SectionHeader an dSectionHeaderButton inside components
